feat: refactor menu items to include reset button (#307)
