+++ src/components/Table/components/TableToolbar/components/ColumnVisibility/columnVisibility.styles.ts
@@ -1,5 +1,4 @@
 import styled from "@emotion/styled";
-import { MenuItem } from "@mui/material";
 import { DropdownMenu } from "../../../../../common/DropdownMenu/dropdownMenu";
 
 export const StyledDropdownMenu = styled(DropdownMenu)`
@@ -11,18 +10,3 @@ export const StyledDropdownMenu = styled(DropdownMenu)`
     }
   }
 `;
-
-export const StyledMenuItem = styled(MenuItem)`
-  &:hover {
-    background-color: transparent;
-
-    .MuiListItemText-root {
-      .MuiListItemText-primary {
-        text-decoration: underline;
-        text-decoration-skip-ink: none;
-        text-underline-color: currentColor;
-        text-underline-position: from-font;
-      }
-    }
-  }
-` as typeof MenuItem;
