+++ src/components/Table/components/TableToolbar/components/ColumnVisibility/components/MenuItems/constants.ts
@@ -1,0 +1,9 @@
+import { ListItemTextProps } from "@mui/material";
+
+export const LIST_ITEM_TEXT_PROPS: Partial<ListItemTextProps> = {
+  primaryTypographyProps: {
+    color: "primary",
+    component: "span",
+    variant: "inherit",
+  },
+};
